backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx>=0.28.1
google-genai==1.14.0
pydantic==2.9.0
aiosqlite==0.20.0
Pillow==10.4.0
